Package tree is no longer deleted on each revision update. It is just updated.

+++ trunk/src/Horn.Core/SourceControl/SvnSourceControl.cs
@@ -50,7 +50,8 @@ namespace Horn.Core.SCM
 
             try
             {
-                packageTree.WorkingDirectory.Delete(true);
+                if (packageTree.Name != PackageTree.RootPackageTreeName)
+                    packageTree.WorkingDirectory.Delete(true);
             }
             catch (IOException)
             {
